FIXED: stopped restricting vectors after init pop.
Shahryar in class today specified that the bounds of the function only applied to the initial population. He does not want us to change the parameter values that have gone outside of the bounds as a result of mutation and crossover. All of the restrictions placed on the vector have been removed besides the restrictions in the contructor used in the creation of the intial population.